refactor operator storage
